Modularize: add global variables and constant value on global.c, move some functions to tools.c & Clean: code alignment, modify annotation
